Set first quote variable to "one".
I need to work out a better way to do this with arrays,
but having the array in the same object as the author
variable doesn't work.

If I don't want to repeat myself for the author I need
to figure out a solution for this, because naming each
quote object with a word isn't gonna fly.