Replace randomArrayFromJSON with extractArrayData.
This is a far better functions in terms of being able to understand
what's actually going on with it. Abstraction is better, too.

Overall I'm happy with the work I've done here. I think the
function needs a different name though, since it's specifially
used for choosing the random objects from the 'background' and
'quote' arrays.