Use a link tag to import fonts from Google.
Doing it how I was with an import at-rule doesn't work with the build
process, since when Normalize is concatenated onto the main stylesheet,
cssmin won't take the import from farther down in the stylesheet and
put it at the top, so it never gets imported.

I think using a link tag is perfectly fine, same number of http requests,
and it's the standard method for importing fonts from Google.