Merge apps into one
This removes the existing project and consolidates all the apps into a
single application.

The Django philosophy is to create lots of little apps that you can plug
together in your project, but this is far too overwhelming as a
beginner.

I've been stuck on figuring out how to distribute all the apps Fireflies
would use, it's a million times easier to have everything in one place.

In the future I fully intend on following the Django way of thinking,
but for the time being I can't continue Fireflies unless I make this
change.